Use watchFile to catch all file updates (i.e. when file is replaced)
fs.watch does not catch all file updates: as it simply watches the
file's inode, it will not detect changes that replace the file and
insert a new inode on the filesystem. It is also not consistent across
different operating systems. We can use fs.watchFile to keep track of
the file and correctly identify all file changes (compare curr.mtime to
prev.mtime to find actual changes as file accesses can trigger events
too). This is more resource-intensive than fs.watch however.